Added SavedPets.js in client/src/pages
